Implements parent 'Entity' class
